chore: Add license, homepage and styling (???)
The project is Licensed under the terms of GNU GPLV3.
Future contributors will now do as they wish as long as it is permitted by the license.

A dedicated homepage is now live (not pretty at all)

Stylings have been added to make the pages pretty to look at.
No functionality have been added but it is pretty as hell!

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>