fix: Use `/view/<slug>` instead of `/view?id={id}`
Change from using id to using the slug.
A preview of the content is shown in the table before a user clicks.
'slug' column has been dropped and every activity requiring a user to acknowledge it.
'slug' is now auto generated and efforts to make it unique with high SEO are going well.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>